+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [[#266](https://github.com/rust-vmm/vm-memory/pull/266)] Derive `Debug` for several
   types that were missing it.
+- [[#274]] Drop `Default` as requirement for `ByteValued`.
 
 ## [v0.13.1]